add tests for a rich section / environment variable
Also add variants for underscores, pipes, multiple spaces, and
underscored-variants of whitespace environment variables. We considered adding
the ability to parse underscored environment variable keys in such a way that
they could "fill in" for not found whitespace keys. However, this created
ambiguity because (1) we have to remove the keys after they are used, lest the
key be used again to create a missing section that we do not want, (2) it is
possible to use both whitespace and underscored variants in section headings so
both, one, or neither could be present, and (3) map ordering is
non-deterministic.

As a result of this, the complexity to implement, and the complexity to correct
this ambiguity, we opted to not implement this functionality for now but leave
tests in place that clarify the nature of our behavior. Right now, no products
using `gcfgenv` have a requirement for whitespace in subsection titles, so we
added a note to advise avoiding such madness. 😄