handle custom types of "kind" Slice with an UnmarshalText method
Custom types of "kind" slice that satisfy the `unmarshaller` interface run into a problem, because `UnmarshalText` intends to unmarshal just a single entry of the slice. As a result, we must handle these cases directly by splitting the environment variable on `,` and calling `UnmarshalText` on each part.

We add such a test struct into the tests to ensure that this behavior works as intended.